django/pyproject.toml
[tool.poetry]
name = "todo-django"
version = "0.1.0"
description = ""
authors = ["mulchus <mulch@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
#python = "^3.12"
requests = "2.31.0"
Django = "^5.2"
django-cors-headers = "3.7.0"
gunicorn = "==21.*"
djangorestframework = "^3.16"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
pytz = "2025.2"
tzdata = "^2025.2"
celery = "^5.5.2"
redis = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
